+++ test/accessor_tests.cc
@@ -669,7 +669,7 @@ namespace detail {
 
 			q.submit([&](handler& cgh) {
 				accessor acc(buff, cgh, celerity::access::fixed(accessible_sr), celerity::write_only, celerity::no_init);
-				cgh.parallel_for<acc_out_of_bounds_kernel<Dims>>(range<Dims>(unit_range), [=](item<Dims>) {
+				cgh.parallel_for<acc_out_of_bounds_kernel<Dims>>(range<Dims>(ones), [=](item<Dims>) {
 					acc[oob_idx_lo] = 0;
 					acc[oob_idx_hi] = 0;
 				});
@@ -677,10 +677,9 @@ namespace detail {
 			q.slow_full_sync();
 		}
 
-		const auto attempted_sr = subrange<3>{id_cast<3>(oob_idx_lo), range_cast<3>(oob_idx_hi - oob_idx_lo + id_cast<Dims>(range<Dims>(unit_range)))};
-		const auto error_message = fmt::format("Out-of-bounds access in kernel 'celerity::detail::acc_out_of_bounds_kernel<{}>' "
-		                                       "detected: Accessor 0 for buffer 0 attempted to access indices "
-		                                       "between {} which are outside of mapped subrange {}",
+		const auto attempted_sr = subrange<3>{id_cast<3>(oob_idx_lo), range_cast<3>(oob_idx_hi - oob_idx_lo + id_cast<Dims>(range<Dims>(ones)))};
+		const auto error_message = fmt::format("Out-of-bounds access in kernel 'celerity::detail::acc_out_of_bounds_kernel<{}>' detected: Accessor 0 for "
+		                                       "buffer 0 attempted to access indices between {} which are outside of mapped subrange {}",
 		    Dims, attempted_sr, subrange_cast<3>(accessible_sr));
 		CHECK_THAT(lc->get_log(), Catch::Matchers::ContainsSubstring(error_message));
 	}