+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2023-??-??
+
+## Added
+
+- Introduce new experimental `for_each_item` utility to iterate over a celerity range (#199)
+
 ## [0.4.1] - 2023-09-08
 
 We recommend using the following SYCL versions with this release: